HC de mediciones, miembro, area y organizacion. Lógica de FE actualizada.

+++ src/main/java/HuellaDeCarbono/CalculoDeHuella/FactoresDeEmision.java
@@ -1,11 +1,9 @@
 package HuellaDeCarbono.CalculoDeHuella;
 
-import HuellaDeCarbono.Repositorios.RepositorioOrganizaciones;
-
 import java.util.HashMap;
 
 public class FactoresDeEmision {
-    private HashMap <String, Float> FE = new HashMap<String, Float>();
+    private HashMap<String, Double> FE = new HashMap<String, Double>();
     private static FactoresDeEmision instance = new FactoresDeEmision();
 
     public static FactoresDeEmision getFactores() {
@@ -14,11 +12,11 @@ public class FactoresDeEmision {
 
     //Se pueden cargar FE con tipos de consumo o con medios de transporte
 
-    public void setFE(String claveDeFE, Float factorDeEmision){
+    public void setFE(String claveDeFE, Double factorDeEmision){
         FE.put(claveDeFE, factorDeEmision);
     }
 
-    public Float getFE(String claveDeFE){
+    public Double getFE(String claveDeFE){
         return FE.get(claveDeFE);
     }
 }
